Split invocations module into submodules.
commit-id:c632cfba
